refactor(settings): extract CalendarSection to sections/
Moves the Calendar tab — macOS Calendar.app permission flow, display
options (events, all-day), and calendar source selection — into its
own file. Keeps the useEffect-driven permission check + calendar fetch
behavior intact.

No behavior changes.
